refactored edit page
