+++ src/components/AddButton.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="text-right mb-3">
+    <b-button variant="outline-info" class="h2 mb-0" v-b-modal.modal-add-form>
+      <b-icon-plus-circle></b-icon-plus-circle>
+    </b-button>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "AddButton"
+};
+</script>
+
+<style></style>
