+++ src/components/EditButton.vue
@@ -1,0 +1,11 @@
+<template>
+  <b-button variant="outline-info" @click="$bvModal.show(`bv-modal-${index}`)"
+    >編集</b-button
+  >
+</template>
+
+<script>
+export default {};
+</script>
+
+<style></style>
