+++ src/main.js
@@ -14,12 +14,7 @@ Vue.use(BootstrapVue);
 Vue.use(IconsPlugin);
 Vue.config.productionTip = false;
 
-axios.defaults.baseURL =
-  "https://firestore.googleapis.com/v1/projects/typing-app-f08b8/databases/(default)/documents";
-//corsのために追加したコード
-// axios.defaults.baseURL = "http://localhost:3000";
-// axios.defaults.headers.post["Content-Type"] = "application/json;charset=utf-8";
-// axios.defaults.headers.post["Access-Control-Allow-Origin"] = "*";
+axios.defaults.baseURL = "https://firestore.googleapis.com/v1/";
 
 store.dispatch("autoLogin").then(() => {
   new Vue({
